organize componenets
