+++ src/Componenets/AmiiboDetails/AmiiboDetails.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-const AmiiboDetails = () => {
+const AmiiboDetails = ({amiibo}) => {
     return (
         <div>
