+++ src/Componenets/App/App.js
@@ -1,9 +1,9 @@
 import React, { useState, useEffect } from "react";
-import Header from "./Header";
-import AmiiboContainer from "./AmiiboContainer";
-import AmiiboDetails from "./AmiiboDetails";
-import UserCollection from "./UserCollection";
-import AboutUs from "./AboutUs";
+import Header from "../Header/Header";
+import AmiiboContainer from "../AmiiboContainer/AmiiboContainer";
+import AmiiboDetails from "../AmiiboDetails/AmiiboDetails";
+import UserCollection from "../UserCollection/UserCollection";
+import AboutUs from "../AboutUs/AboutUs";
 import { Route, Switch} from 'react-router-dom';
 
 const App = () => {
