+++ src/Componenets/UserCollection/UserCollection.css
@@ -1,5 +1,5 @@
 .user-amiibo-collection-container{
-    background-image: url("Shelf.png");
+    background-image: url("../Images/Shelf.png");
     background-size: cover;
     background-repeat: no-repeat;
     background-attachment: fixed;
